nep-78-refresh-btn: no internet and office message is fixed
